+++ CHANGELOG.md
@@ -81,6 +81,12 @@
 - Fixed a bug where a Unity package with no dependencies field in its `package.json` would cause code generation to throw exceptions.
 - Fixed a bug where protocol logging would crash Linux workers.
 
+### Added
+- Added a new Query-based interest helper module, `com.improbable.gdk.querybasedinteresthelper`.
+    - `Improbable.Gdk.QueryBasedInterestHelper.InterestHelper` provides functionality to ergonomically add, replace and clear queries from an Interest component.
+    - `Improbable.Gdk.QueryBasedInterestHelper.InterestQuery` enables simpler construction of interest queries.
+    - `Improbable.Gdk.QueryBasedInterestHelper.Constraint` contains static methods to construct the QueryConstraint objects which used to define interest queries.
+
 ## `0.1.4` - 2019-01-28
 
 ### Added